requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
python-jose[cryptography]
passlib[bcrypt]
pandas
python-multipart
python-dotenv
psycopg2-binary